+++ src/tools/create_mulmo_script_interactively.ts
@@ -210,7 +210,7 @@ const scrapeWebContent = async (urls: string[], cacheDirPath: string) => {
   return `\n\n${prefixPrompt}\n${result?.sourceText.text}`;
 };
 
-export const createMulmoScriptWithInteractive = async ({ outDirPath, cacheDirPath, filename, templateName, urls }: ScriptingParams) => {
+export const createMulmoScriptInteractively = async ({ outDirPath, cacheDirPath, filename, templateName, urls }: ScriptingParams) => {
   mkdir(outDirPath);
   // if urls is not empty, scrape web content and reference it in the prompt
   const webContentPrompt = urls.length > 0 ? await scrapeWebContent(urls, cacheDirPath) : "";
